fix BLE decoding to use a signed short for proper negative handling
The sample code that was used had used an unsigned long for the "value" variable.  This leads to problems - specifically temperature calculations - when the temperature drops below 0C and become negative.  The code was only using 16-bits within the long (32-bits) and as it was unsigned it was not handling negative numbers properly.